update UI page, fix bugs delete feedback

See merge request !86

+++ websites/api_admin/views.py
@@ -249,7 +249,7 @@ class PromotionStatistic(APIView):
 
                 result['promotion'] = admin_serializers.PromotionSerializer(
                     promotion_detail, many=False).data
-                result['gift-user'] = admin_serializers.GiftSerializer(gift_list, many=True).data
+                result['gift_user'] = admin_serializers.GiftSerializer(gift_list, many=True).data
 
                 result['count_user_total'] = user_promotion_list.count()
                 result['count_user_device'] = promotion_user_id_list.exclude(device_id__isnull=True).count()
@@ -2061,16 +2061,21 @@ class UserListView(APIView):
     def delete(self, request, format=None):
         print "METHOD DELETE"
         try:
-            # Get list user id to delete
-            user_id = self.request.data.get('user_id', None)
-            print "User List Id", user_id
+            # Get role_id
+            role_id = self.request.user.role_id
+            # Check role_id
+            if role_id == 1:
+                # Get list user id to delete
+                user_id = self.request.data.get('user_id', None)
+                print "User List Id", user_id
 
-            # Check list user id
-            if user_id:
-                User.objects.filter(pk__in=user_id).delete()
-                return Response({"code": 200, "message": _("success"), "fields": ""}, status=200)
-            return Response({"code": 400, "message": "List ID Not found ", "fields": ""}, status=400)
-
+                # Check list user id
+                if user_id:
+                    User.objects.filter(pk__in=user_id).delete()
+                    return Response({"code": 200, "message": _("success"), "fields": ""}, status=200)
+                return Response({"code": 400, "message": "List ID Not found ", "fields": ""}, status=400)
+            else:
+                return Response({"code": 405, "message": _("Just System Admin accept delete"), "fields": ""}, status=405)
         except Exception, e:
             print e
             error = {"code": 500, "message": _("Internal Server Error"), "fields": ""}
